+++ src/app.ts
@@ -3,9 +3,12 @@ import { parseTemplate } from './elements/utils'
 import './elements/audio-player'
 import './elements/description-display'
 import './elements/sound-menu'
+import './elements/number-input'
 import { type SoundMenu } from './elements/sound-menu'
-import { type SoundEntry, type SoundState, type AudioPlayer } from './elements/audio-player'
+import { type AudioPlayer } from './elements/audio-player'
 import { type DescriptionDisplay } from './elements/description-display'
+
+import { soundStateFromURL, playerStateFromURL } from './elements/url-utils'
 
 import template from './templates/app.html'
 import './style/app.css'
@@ -20,11 +23,10 @@ export class App extends HTMLElement {
   constructor () {
     super()
     this.appendChild(nodes.cloneNode(true))
+
     this.descriptionDisplay = this.querySelector('description-display') as DescriptionDisplay
     this.audioPlayer = this.querySelector('audio-player') as AudioPlayer
     this.audioPlayer.initialize({
-      addSoundToURL: App.addSoundToURL,
-      removeSoundFromURL: App.removeSoundFromURL,
       findSound: (soundName) => this.soundMenu.sounds[soundName]
     })
     this.soundMenu = this.querySelector('sound-menu') as SoundMenu
@@ -42,11 +44,15 @@ export class App extends HTMLElement {
       }
     })
     this.connectEvents()
+    this.audioPlayer.loadPlayerState(playerStateFromURL())
     this.soundMenu.soundsAdded
       .then(() => {
-        this.audioPlayer.loadState(soundStateFromURL())
+        this.audioPlayer.loadSoundState(soundStateFromURL())
       })
       .catch((error) => { throw error })
+
+    // To enable blurring input elements by clicking anywhere
+    this.tabIndex = 0
   }
 
   connectEvents (): void {
@@ -70,78 +76,27 @@ export class App extends HTMLElement {
   }
 
   keyboardInteraction (event: KeyboardEvent): void {
-    if (['ArrowRight', 'ArrowLeft'].includes(event.code)) {
+    if (
+      ['ArrowRight', 'ArrowLeft'].includes(event.code) &&
+        (event.target as HTMLElement).tagName !== 'INPUT'
+    ) {
       event.preventDefault()
       if (this.audioPlayer.started == null) {
         if (event.code === 'ArrowRight') {
-          const newStart = this.audioPlayer.start + 0.1
+          const newStart = this.audioPlayer.position + 0.1
           if (newStart < this.audioPlayer.duration) {
-            this.audioPlayer.start = newStart
+            this.audioPlayer.position = newStart
           }
         }
         if (event.code === 'ArrowLeft') {
-          const newStart = this.audioPlayer.start - 0.1
+          const newStart = this.audioPlayer.position - 0.1
           if (newStart >= -1e-10) {
-            this.audioPlayer.start = newStart
+            this.audioPlayer.position = newStart
           }
         }
       }
     }
   }
-
-  static addSoundToURL (refNum: number, soundEntry: SoundEntry): void {
-    const searchParams = new URLSearchParams(window.location.search)
-    searchParams.set(`sound${refNum}`, soundEntry.soundName)
-    searchParams.set(`track${refNum}`, `${soundEntry.trackNum}`)
-    searchParams.set(`start${refNum}`, `${soundEntry.start}`)
-    const newURL = window.location.pathname + '?' + searchParams.toString()
-    history.replaceState(null, '', newURL)
-  }
-
-  static removeSoundFromURL (refNum: number): void {
-    const searchParams = new URLSearchParams(window.location.search)
-    searchParams.delete(`sound${refNum}`)
-    searchParams.delete(`track${refNum}`)
-    searchParams.delete(`start${refNum}`)
-    const newURL = window.location.pathname + '?' + searchParams.toString()
-    history.replaceState(null, '', newURL)
-  }
-}
-
-function soundStateFromURL (): SoundState {
-  const searchParams = new URLSearchParams(window.location.search)
-  const soundState = new Map()
-  for (const [key, value] of searchParams.entries()) {
-    let refNum
-    let paramName
-    let paramValue
-    refNum = key.split('sound')[1]
-    if (refNum != null) {
-      paramName = 'soundName'
-      paramValue = value
-    } else {
-      refNum = key.split('track')[1]
-      if (refNum != null) {
-        paramName = 'trackNum'
-        paramValue = Number(value)
-      } else {
-        refNum = key.split('start')[1]
-        paramName = 'start'
-        paramValue = Number(value)
-      }
-    }
-    if (refNum != null) {
-      refNum = Number(refNum)
-      if (soundState.get(refNum) != null) {
-        soundState.get(refNum)[paramName] = paramValue
-      } else {
-        const state: Record<string, number | string> = {}
-        state[paramName] = paramValue
-        soundState.set(refNum, state)
-      }
-    }
-  }
-  return soundState
 }
 
 customElements.define('app-', App)